requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pg8000
psycopg2-binary
requests
python-dateutil
httpx
PyJWT
cachetools
unidecode
onnxruntime-gpu
opencv-python
matplotlib
networkx
pytest
shapely
pyclipper
pdf2image
python-multipart